feat: #18 Utilsモジュールへ分割
